reference/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
pyogrio>=0.7.0
fiona>=1.10.0
shapely>=2.0.0
pyproj>=3.6.0
geopandas>=0.14.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
ezdxf>=1.0.0
mcp>=1.0.0
slowapi>=0.1.9
